+++ app/page.tsx
@@ -1,5 +1,5 @@
-import Link from "next/link";
 
+import Link from "next/link";
 export default function Home() {
   return (
     <section className="w-full h-screen grid place-items-center bg-dark">
