--- lib/utils.ts
@@ -1,2 +1,0 @@
-// add common utility functions here that will be user throughout the project
-
